numeric cleanup
