+++ app/packages/state/src/recoil/pathData/lightningNumeric.ts
@@ -16,10 +16,7 @@ export const lightningNumericResults = selectorFamily({
   get:
     (path: string) =>
     ({ get }) => {
-      const grouped = get(numericFields(path));
-      const index = grouped.findIndex((p) => p === path);
-      const results = get(lightningQuery(grouped.map((p) => ({ path: p }))));
-      const data = results[index];
+      const [data] = get(lightningQuery([{ path }]));
 
       if (data.__typename === "DateLightningResult") {
         return {
